requirements.txt
streamlit==1.49.1
pandas==2.3.2
numpy==1.26.4
boto3==1.34.34
botocore==1.34.34
aiobotocore==2.11.2
langfuse==3.5.0
python-dotenv==1.1.1
itables==2.5.2
fsspec==2024.2.0
s3fs==2024.2.0
matplotlib==3.10.6
seaborn==0.13.2
scikit-learn==1.5.0
joblib==1.4.2
google-generativeai==0.8.3